옷 추가/수정 화면 키보드 개선
- 키보드 위 툴바(이전/다음/완료) 추가로 필드 간 이동 지원
- 별점 0.5단계 입력 지원 (rating: Int → Double 타입 변경)
- 탭 영역 확장 및 키보드 외부 터치 시 닫힘 처리

+++ Weather_closet/Domain/Entities/Closet/ClothingEntity.swift
@@ -10,7 +10,7 @@ struct ClothingEntity: Identifiable {
     var color: String
     var size: ClothingSize
     var alterationHistory: [AlterationRecord]
-    var rating: Int
+    var rating: Double
     var review: String
     var wearCount: Int
     var purchaseDate: Date?
